Extract CMS content from Confirmation Page into component
